Update Main.ino
Perfect sketch for the ArduinoSpotifySample by witnessmenow edited by Acorn221 and updated by Sebbl2.
I had many issues with that Sketch starting with the wrong board chosen. Also there was wrong set backslash, so it wouldn't compile. esp32 dev board was fine either with the Spotify-api-sample or the LilyGo Example one or two. Not working at all with the SpotifyDisplayLILYGO Sketch. Compiling was OK (after fixing the \) but it never starts to run.
After choosing the ESP32Pico Kit,it worked, but with a lot of fragments on the display and without Logo and barcode. 
I switched to GxGDEM0213B74/GxGDEM0213B74.h alternatively GxGDEH0213B73/GxGDEH0213B73.h both worked like charm, no Logo nor Barcode.
there where some http errors on the serial monitor so I decided to kill the reverence to the certificate. And it worked flawlessly. Thank you Arcorn221 for these two days. Ive learned a lot. Hopefully it helps you.